auth-service/requirements.txt
flask
flask-cors
boto3
python-jose[cryptography]
python-dotenv
requests